feat: connect search presenter to load device from API
This won't do anything for now. Even though everything (except
the details view) is connected now, the wrapper needs to be
added to the project before I start making API calls to the
NLM database. Once the wrapper is added and the details
view is fully built, this app should be pretty much done.

Closes: #7