feat: implement details presenter methods
NOTE: Like the Search Presenter some of the methods
still need the wrapper to be able to fully work.
Once the wrapper is added, everything should work.
The way this was implemented means we'll have to
change how activities go to the Details Activity,
but that'll be easy to fix.

Closes #10