implement new UDIWrapper when loading device from API
The API now requires an API Key. This is now in values/strings.xml.

To get that string, the LoadDeviceFromApi class needs a context, which is provided in its constructor.

The LoadDeviceFromApi class now creates a new builder when it's constructed, then builds the UDIWrapper when one of its methods is called.

Since the LoadDeviceFromApi has a private UDIWrapper and UDIWrapper.Builder, it doesn't need to recreate those classes every time a search is kicked off. It can just alter the search and return the relevant data.