implement the required changes. this is a rough draft
noDevice will now set the visibility to none, and clear the adapter on the listview.

setDevices checks for the searchChangedFlag. If the search was changed, it clears the adapter, if not, it adds items to the adapter and updates the listview.

The search presenter is doing some weird stuff to manage state and appropriately react to the end of the list, update the devices, etc.