keep track of the UDIWrapper better
Now the LoadDeviceFromApi implementation will check that the skip, search value and search property have all changed before it alters the search. This keeps the model from making two API calls.

Instead of checking if the device exists in one call, and then making a call to get devices (which is unnecessary with this library), the model checks if the device exists and then immediately gets the associated devices.